Fixed a bug where we were importing from a worker file and thus inadvertently copying its Pyodide initialisation on the main thread.  Would explain some of the test slowdown.
